backend/requirements-dev.txt
fastapi[standard]==0.115.8
glom==24.11.0
mock==5.1.0
pyright==1.1.401
pytest==8.3.4
pytest-asyncio==0.26.0
pytest-httpx==0.35.0
types-PyYAML==6.0.12.20241230
types-requests==2.32.0.20241016